build(ci,python): raise ruff to 0.16 and settle its new default rules
- nothing above tools/, tests/, benchmarks/ and examples/ carries a ruff
  configuration, so those files run on whatever rule set the pinned ruff
  defaults to; 0.16 widened that default, which is why the lock could not
  be followed when it last moved
- the workflow pins and the lock go to 0.16.8 together, so make lint and
  both CI jobs run the linter the lock installs again
- most of the tree change is mechanical: the decorative shebangs go, since
  no python file is executable and nothing invokes one by path; regex flag
  aliases and printf formatting are spelled out; dict() calls and explicit
  string concatenation become literals; mutable class attributes are
  annotated
- every subprocess.run gains an explicit check=False, because all of them
  already read returncode and raise their own message, and check=True would
  replace that message with a CalledProcessError
- closures a loop consumes inside its own iteration bind their loop
  variables as defaults, matching the idiom the coordinate search already
  used for the two it bound
- where a new rule argues against a deliberate choice the site carries a
  noqa and its reason: the NaN self-comparisons, the blind excepts that
  report rather than re-raise, the single error class a document validator
  keeps and its tests pin, and the two percent-templates whose braces
  belong to the language they generate
- the conditions the auto-fix folded into one long line stay folded but are
  re-wrapped, and the coverage chain in the parity comparison stays two
  statements so it still reads as a pair with the one below it

+++ tools/voicematch/corpus.py
@@ -32,7 +32,6 @@ from dataclasses import dataclass, field
 from pathlib import Path
 
 import numpy as np
-
 from capture import (
     RIG_BAKED,
     RIG_NONE,
@@ -540,10 +539,10 @@ def corpus_oracle(corpus: Corpus, pattern: Pattern, sr: int) -> np.ndarray:
             f"resampling a reference"
         )
     last = max(pattern.notes, key=lambda n: n.start)
-    total = int(round(
-        (last.start + corpus.slot_for(corpus.capture_slot(last.note), last.velocity)) * sr))
+    total = round(
+        (last.start + corpus.slot_for(corpus.capture_slot(last.note), last.velocity)) * sr)
     out: np.ndarray | None = None
-    skip = int(round(corpus.preroll_s * sr))
+    skip = round(corpus.preroll_s * sr)
     # The pattern is written in the model's numbering; the renders are filed
     # under the notes the reference was struck on. See `Corpus.capture_slot`.
     for note in pattern.notes:
@@ -559,9 +558,9 @@ def corpus_oracle(corpus: Corpus, pattern: Pattern, sr: int) -> np.ndarray:
         if out is None:
             out = np.zeros((total, audio.shape[1]), dtype=np.float64)
         seg = audio[skip:]
-        start = int(round(note.start * sr))
+        start = round(note.start * sr)
         slot = corpus.slot_for(slot_note, note.velocity)
-        room = min(len(seg), total - start, int(round(slot * sr)))
+        room = min(len(seg), total - start, round(slot * sr))
         if room > 0:
             out[start : start + room] += seg[:room]
     if out is None:
